Fix upload step
Commit f728f884e changed the archive built charm step. After that
commit two tasks both register the charm_archived variable. Which
task is run depends on the `when` clause. The problem with this is
that the second task always overwrites the contents of `charm_archived`
even if the tasks `when` condition is not met. If the second task is
skipped then `charm_archived` is a boolean set to False. If the task
runs `charm_archived` is a dict containing the return code of the tar
command. If the second task is skipped then a subsequent task fails
because it assume `charm_archived` to be a dict *1

This patch changes the archive tasks to register different variables
and then updates the upload task to check both variables.

There is also a drive-by fix to the charmcraft tar step to stop
it including the tar ball in the tar ball.

*1 https://openstack-ci-reports.ubuntu.com/artifacts/45f/805079/1/check/charm-build/45f050d/job-output.json