Tar charmcraft build charms in temp dir
The charmcraft build outputs to the current directory so the tar
is of the current directory too unlike reactive charms which are
built in a subdirectory and the tar ball created from a few directories
up.

Although the charm being built should be excluded from the archive
sometimes `tar: .: file changed as we read it ` errors are seen.
To guard against this build the tar ball in a seperate directory.