Only upload built .charm file for charmcraft charms
9727ca97ef was reverted due to a fatal error when building
reactive charms:

`
2021-10-05 10:42:04.982545 | xenial-medium |   "msg": "The conditional check 'needs_charm_build and fetch_charm.stdout.find(\"successfully fetched built\") == -1' failed. The error was: error while evaluating conditional (needs_charm_build and fetch_charm.stdout.find(\"successfully fetched built\") == -1): 'dict object' has no attribute 'stdout'\n\nThe error appears to be in '/var/lib/zuul/builds/e81c9accd06b47dd836b36093e7d504c/trusted/project_0/github.com/openstack-charmers/zosci-config/roles/charm-build/tasks/main.yaml': line 41, column 3, but may\nbe elsewhere in the file depending on the exact syntax problem.\n\nThe offending line appears to be:\n\n#      position of that, as -1 is the value for a not-found in this case.\n- name: build charm\n  ^ here\n"
`

https://openstack-ci-reports.ubuntu.com/artifacts/e81/812464/1/check/charm-build/e81c9ac/job-output.txt

This appears to be because both the reactive and the charmcraft
fetch jobs both registered `fetch_charm`. The reactive one is
defined first and run first. The charmcraft job is run next and
the value of `fetch_charm` is overwritten even if the jobs
conditional is not met. This means when the subsequent job
tries to examine fetch_charm.stdout it finds it non-existant.

This PR contains 9727ca97ef plus defining different fetch_charm
vars for the reactive and charmcraft builds. This inturn
requires an update to the `build charm` job so that it checks
both fetch parameters.