Run rudimentary checks on downloaded file
Before building a charm zosci attempts to download it from swift.
If it succeeds then the build is skipped. However the download
always succeeds, if the artifact isn't present in swift the
'downloaded' charm just contains 'NoSuchKey'

`
$ curl -o /tmp/foo.charm  http://10.245.161.162:80/swift/v1/zuul-built-charms/foo.charm
$ cat /tmp/foo.charm
NoSuchKey
`

For reactive charms the charm is downloaded and then untar'd but
the untar fails which stop job from echoing the success message.
For charmcraft charms there is no untar step so the job appearts
to have correctly downloaded the charm.

This PR adds a very rudimentary check on the downloaded file
to ascertain whether the charm was downloaded correctly.