Fix address used for name server
The current configuration makes use of a behavior of OpenStack
Neutron with ML2/OVS which was to provide a dnsmasq-based DNS
forwarder on the .2 address of project networks.

This behavior is not retained with ML2/OVN and we need to use the
CI infrastructure upstream DNS server directly instead.

Cloud-internal instance name DNS resolution still works because
DNS responses will be intercepted by Open vSwitch and the
ovn-controller when it knows it can provide an answer to the
query being sent to the real server.