Temporarilly set juju model-default userdata
To help debug an intermittent and not reproducible issue with DNS
resolution seen in gate, have juju add cloud-init userdata that
enables debug logging for systemd-resolved.

Note that the juju model-config key `cloudinit-userdata` must be
fed in through a file and cannot be set on command line nor fed
in through the Zaza model settings environment variables.