Ensure that simple function jobs have required dependencies
Due to a misunderstanding on my part, the 'simple' jobs (such as
focal-wallaby) only had the tox-pyVV as a job dependencies as I though
the parent job's dependencies were merged.  However, this is not the
case. Declaring a 'dependencies' key in a job replaces the parent job's
'dependencies' key; in this case it means that the func job won't wait
for charm-build of the lint.  This patch corrects that.