Allow osci.yaml to control the charmcraft build version
Allow the osci.yaml to include a var `charmcraft_channel` that can be
optionally used to override the default charmcraft channel that the snap
installs from.  This is to allow different branches to control which
version of charmcraft that they should use.  The default version is
fixed at "1.5/stable" which is used for all the backport branches.