Upload multiple charmcraft charms
Charmcraft allows building multiple charm artifacts from a single
repository, pivoting on Ubuntu series and target architectures.

To support this, instead of uploading a single named .charm
artifact, we archive up all .charm files in the work directory
and upload that archive to Swift.

The change re-uses the archive type and file naming from the
old reactive build job, and as such some of the handling becomes
equal for both charm types.