Update charm-build job to run on jammy
Previously the `charm-build` job had to run on a Xenial image in
order for `charm-tools` to produce compatible reactive charms.

Since then all charms have been updated to make use of `charmcraft`
to execute the build.

Charmcraft executes the build in LXD containers using the series
specified in the charmcraft.yaml file, and as such there is no
longer a need to keep the outer image on Xenial.