Adapt use of tox_extra_args to tox 4.x
The outer tox which is used to start for example the `func-target`
environment, is installed by the Zuul role named `ensure-tox`.

This role pulls tox from PyPI and that means that test executors
running modern operating systems will now pull in tox 4.x.

Tox 4.x requires the use of `--` to separate tox and environment
{posargs}, so this change adds that to the job definitions.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>