Handle tox_config_file for functional tests on source charms
During the transition to charmcraft builds, a hack was added to
the charms outer tox.ini in order to keep functional tests
working [0].

A recent update to tox changed the argument parsing which broke
that hack, and it has become apparent that we ought to fix this
properly.

Add a new role that checks for the presence of src/tox.ini and
conditionally set the `tox_config_file` variable appropriately.

The new role is called as part of the juju playbook, which is
used by functional test jobs only.

0: https://github.com/openstack-charmers/release-tools/blob/874e86cebfb0/global/source-zaza/tox.ini#L87-L97