Update charm-unit-jobs and charm-functional-jobs for antelope
Instead of creating an charm-antelope-unit-jobs and
charm-antelope-functional-jobs template, we switch to using the
charm-unit-jobs and charm-functional-jobs templates with their
'branches' specyfing the jobs to run. This allows for managing
jobs in a central location.

This is similar to changes being made to openstack-zuul-jobs,
such as in [1], and you'll notice I've borrowed the inline
comments from gmann.

[1]
https://review.opendev.org/c/openstack/openstack-zuul-jobs/+/873907