Rewrite jinja2 snippet to determine if juju snap should be classic
The commit 16b1561 introduced a dynamic evaluation to set the `classic`
property to true/false when installing the juju snap, although that line
produces a "False" which Ansible is evaluating to true.

This change takes a different approach dropping the use of an inline-if
which was producing a string value, now it's being used a conditional
expression to force the evaluation as a boolean.