Handle constraints from a URL
The tox job[0] accepts the tox_constraints_file variable, although its
implementation checks that the string references a locally available
file[1]. This prevents from passing a url to pip.

This change reads the content from a url passed via pip_constraints_url
and renders it in a temporary file that gets referenced to be used by
tox.

[0] https://zuul-ci.org/docs/zuul-jobs/latest/python-jobs.html#jobvar-tox.tox_constraints_file
[1] https://opendev.org/zuul/zuul-jobs/src/branch/master/roles/tox/tasks/main.yaml