Set constraints for bobcat jobs
The *-bobcat jobs are run by master branch AND stable/2023.2, while the
former would be OK to be moved to use juju-3.x, the latter is not.

Without this change, bobcat jobs running in master will install
juju-2.9, but zaza from master making it install libjuju-3.x. This
change introduces the constraint which will be honored by master and
discarded by stable/2023.2 although that stable branch installs zaza
from stable/bobcat which has libjuju pinned to `<3.0`.