Force download of constraints files
Ansible's get_url module uses the last modified time of an existing
file on disk to only download files that have potentially changed.
The handle-constraints-url role will create a temporary temporary
file to store the constraints file, which leaves the last modified
time as extremely recent. This causes the remote webserver to respond
with an HTTP 304 (content not modified) and the constraints do not
get properly set.

This change addresses this by adding force: true to the get_url
module which will always download the specified file and not use
the file's last modified timestamp.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>