Modify func-test-pr to work with globbed filenames
The stable/X branches of OpenStack use python version specific
requirements files, in the form `requirements-pyX.txt`. The way that
func-test-pr was implemented limited scanning/updating files to those in
the form `requirements.txt`, etc.

This patch introduces globbing support, and then in the ansible
playbook, globbing for `requirements*.txt` rather to enable multiple
files to be updated via the func-test-pr.